Move util methods to class, minor changes
