+++ ctb replay editor/PlayField.cs
@@ -6,7 +6,6 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
 using osu_database_reader.Components.HitObjects;
-using ReplayAPI;
 using Un4seen.Bass;
 using Keys = ReplayAPI.Keys;
 
@@ -15,7 +14,6 @@ namespace ctb_replay_editor {
         public float ApproachRateInMS = 300f;
 
         public float CharPos = 256f;
-        public float CircleSize = 0.5f;
 
         public bool GoingLeft;
         public bool IsDash;
@@ -37,41 +35,17 @@ namespace ctb_replay_editor {
             Control.FromHandle(Window.Handle).VisibleChanged += (sender, args) => ((Control) sender).Visible = false;
         }
 
-        public static double ApplyModsToDifficulty(double difficulty, double hardRockFactor, Mods mods) {
-            if (mods.HasFlag(Mods.Easy))
-                difficulty = Math.Max(0, difficulty/2);
-            if (mods.HasFlag(Mods.HardRock))
-                difficulty = Math.Min(10, difficulty*hardRockFactor);
-
-            return difficulty;
-        }
-        
-        public double AdjustDifficulty() {
-            return (ApplyModsToDifficulty(CircleSize, 1.3, Mods.None) - 5)/5;
-        }
-
-        public float GetCatcherWidth() {
-            //0,4
-            float SpriteDisplaySize = (float) (512/8f*(1f - 0.7f*AdjustDifficulty())); //== 7,68
-            //
-            float SpriteRatio = SpriteDisplaySize/128;
-            //0,06
-            float catcherWidth = 305/1*SpriteRatio*0.7f;
-            //12,81
-            return catcherWidth;
-        }
-
         private Texture2D TextureFromFile(string path, int width, int height) {
             try {
                 return Texture2D.FromStream(GraphicsDevice, new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.Read), width, height, true);
             } catch (Exception) {
-                return TextureFromColor(Color.White, 20, 20);
+                return TextureFromColor(Color.White, width, height);
             }
         }
 
         private Texture2D TextureFromColor(Color color, int w = 1, int h = 1) {
-            Texture2D texture = new Texture2D(GraphicsDevice, w, h);
-            Color[] data = new Color[w*h];
+            var texture = new Texture2D(GraphicsDevice, w, h);
+            var data = new Color[w*h];
             for (int i = 0; i < w*h; i++)
                 data[i] = color;
             texture.SetData(data);
